fixup the build error .text not fit in region iram1_0_seg
The old libdriver.a and libmqtt.a are not generated with the flag
'-DICACHE_FLASH'. This cause the ICACHE_FLASH_ATTR invalid and place
all the functions into .iram1_0_seg section.

The size of the .iram1_0_seg section is 0x8000 in all ld scripts
ld/eagle.app.v6.* of the esp sdk